# src/main/resources/openapi-spec.yaml
openapi: 3.0.1
info:
  title: Proba auth microservice
  description: Authentication service for Proba system
  contact:
    name: Alexandr Cherentsov
    email: sanhugo4u@hotmail.com
  version: 0.1.0
servers:
  - url: https://localhost:8081

tags:
  - name: Registration
    description: Ручки регистрации новых пользователей / Endpoints for registration new users
  - name: Authentication
    description: Ручки входа в систему и выдачи токена / Endpoints for generating tokens
  - name: Superadmin
    description: Возможности супер-администратора / Opportunities for superadmin
  - name: Internal_auth
    description: Ручки внутреннего пользования (не запрашиваемые напрямую пользователем) / Internal API
  - name: Password
    description: Ручки сброса пароля / Endpoints for changing password

paths:
  /api/auth/register:
    post:
      tags:
        - Registration
      summary: Регистрация нового пользователя
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationDTO"
      responses:
        201:
          description: Created
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/role/add:
    post:
      tags:
        - Superadmin
      summary: Добавление роли пользователю
      operationId: addRoleToUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAndRoleDTO"
      responses:
        200:
          description: Success
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Авторизация и получение токенов и ID сессии
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDTO"
      responses:
        200:
          description: Вход в систему
          headers:
            Set-Cookie:
              description: Set both tokens to Redis cache, gives Session-ID to HttpOnly Cookie
              schema:
                type: string
                example: "_Host-session_id=etd...; Path=/ HttpOnly; Secure SameSite=Strict; Max-Age=604800"
        404:
          $ref: "#/components/responses/ErrorResponse"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Выход из системы
      operationId: logoutUser
      security: []
      parameters:
        - name: session_id
          in: cookie
          required: true
          description: session ID
          schema:
            type: string
      responses:
        200:
          description: Successfully logged out
          headers:
            Set-Cookie:
              description: Cookie Deletion instructions
              schema:
                type: string
                example: "session_id=; Path=/; Max-Age=0; HttpOnly; Secure; SameSite=Strict"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /internal/auth/refresh:
    post:
      tags:
        - Internal_auth
      summary: Вызов для обновления токенов
      operationId: updateTokens
      security: []
      parameters:
        - name: session_id
          in: cookie
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        200:
          description: Токены обновлены успешно / Tokens updated successfully
          headers:
            Set-Cookie:
              description: Set both tokens to Redis cache and set Session-ID to
              schema:
                type: string
                example: "_Host-session-ID=etd...; Path=/ HttpOnly; Secure SameSite=Strict; Max-Age=604800"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/clear_password:
    post:
      tags:
        - Password
      summary: Сброс пароля (запрос кода на почту)
      operationId: clearPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDTO"
      responses:
        200:
          description: Success
        404:
          $ref: "#/components/responses/ErrorResponse"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/validate:
    post:
      tags:
        - Password
      summary: Принимает код подтверждения для сброса пароля
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeAndEmailDTO"
      responses:
        200:
          description: Code is sent
        404:
          $ref: "#/components/responses/ErrorResponse"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/change_password:
    post:
      tags:
        - Password
      summary: Смена пароля (принимает почту и новый пароль)
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewPasswordDTO"
      responses:
        200:
          description: Success
        404:
          $ref: "#/components/responses/ErrorResponse"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/confirm_registration:
    post:
      tags:
        - Registration
      summary: Эндпоинт для подтверждения регистрации (нужен код для завершения регистрации)
      operationId: confirmRegistration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalCodeDTO"
      responses:
        200:
          description: Success
        404:
          $ref: "#/components/responses/ErrorResponse"
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/resend:
    post:
      tags:
        - Registration
      summary: Endpoint for generating new confirmation token
      operationId: resendConfirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailDTO"
      responses:
        202:
          description: Confirmation code is sent
        500:
          $ref: "#/components/responses/ErrorResponse"
  /api/auth/role/remove:
    post:
      tags:
        - Superadmin
      summary: Отзыв роли пользователя
      operationId: removeRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAndRoleDTO"
      responses:
        200:
          description: Success
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/ErrorResponse"

components:
  schemas:
    Token:
      type: object
      description: Access-токен
      properties:
        token:
          type: string
      required:
        - token
    ApprovalCodeDTO:
      type: object
      description: DTO с кодом подтверждения регистрации
      properties:
        code:
          type: string
          minLength: 8
      required:
        - code

    CodeAndEmailDTO:
      type: object
      description: DTO с почтой и кодом подтверждения сброса пароля
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
        email:
          type: string
          format: email
          minLength: 4
      required:
        - code
        - email
    EmailDTO:
      type: object
      description: DTO с почтовым адресом пользователя, для которого нужно сбросить пароль
      properties:
        user_email:
          type: string
          format: email
          minLength: 4
      required:
        - user_email
    NewPasswordDTO:
      type: object
      description: DTO with account email and new password
      properties:
        user_email:
          type: string
          format: email
          minLength: 4
        new_password:
          type: string
          minLength: 8
    UserAndRoleDTO:
      type: object
      description: DTO с пользователем и ролью (для удаления/добавления роли)
      properties:
        user_data:
          type: string
          minLength: 3
        role:
          type: string
          format: role
      required:
        - user_data
        - role
    UserRegistrationDTO:
      type: object
      description: DTO с регистрационными данными пользователя
      properties:
        login:
          type: string
          nullable: false
          minLength: 1
        email:
          type: string
          nullable: false
          minLength: 5
        password:
          type: string
          nullable: false
          minLength: 9
      required:
        - login
        - email
        - password
    UserLoginDTO:
      type: object
      description: DTO для входа пользователя в систему и получения JWT-токена
      properties:
        login:
          type: string
          nullable: false
          minLength: 1
        password:
          type: string
          nullable: false
          minLength: 8
      required:
        - login
        - password
    Error:
      type: object
      description: Error DTO
      properties:
        description:
          type: string
          description: Описание ошибки
        time:
          type: string
          format: date-time
          description: время возникновения ошибки
      required:
        - description
        - time
  responses:
    Forbidden:
      description: Запрещенный вызов
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Success:
      description: Успешно
      content:
        application/json:
          schema:
            type: string
    ErrorResponse:
      description: DTO для ошибок
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"